login

+++ app/Http/Controllers/ProductController.php
@@ -14,6 +14,15 @@ class ProductController extends Controller
         return view('products.index', ['products' => $products]);
     }
 
+    public function showMyProducts()
+    {
+        $products = [];
+        if (auth()->check()) {
+            $products = auth()->user()->userProducts()->get();
+        }
+        return view('products.myProduct', ['products' => $products]);
+    }
+
     public function create()
     {
         return view('products.create');
@@ -22,12 +31,16 @@ class ProductController extends Controller
     public function store(ProductRequest $request)
     {
         $data = $request->validated();
-        $newProduct = Product::create($data);
+        $data['user_id'] = auth()->id();
+        Product::create($data);
         return redirect()->route('product.index');
     }
 
     public function edit(Product $product)
     {
+        if (auth()->user()->id !== $product['user_id']) {
+            return redirect()->back();
+        }
         return view('products.edit', ['product' => $product]);
     }
 
@@ -40,6 +53,9 @@ class ProductController extends Controller
 
     public function destroy(Product $product)
     {
+        if (auth()->user()->id !== $product['user_id']) {
+            return redirect()->back();
+        }
         $product->delete();
         return redirect()->route('product.index')->with('success', 'Product updated successfully');
     }
